fix: move util functions with single ref
@W-11748691@
