+++ messages/package_version_create.md
@@ -73,3 +73,23 @@ Multiple errors occurred:
 # invalidDaysNumber
 
 Provide a valid positive number for %s.
+
+# errorAncestorNoneNotAllowed
+
+Can’t create package version because you didn’t specify a package ancestor. Set the ancestor version to %s, and try creating the package version. You can also specify --skipancestorcheck to override the ancestry requirement.
+
+# errorAncestorNotHighest
+
+Can’t create package version. The ancestor version [%s] you specified isn’t the highest released package version. Set the ancestor version to %s, and try creating the package version again. You can also specify --skipancestorcheck to override the ancestry requirement.
+
+# invalidPackageTypeMessage
+
+Invalid package type
+
+# errorInvalidBuildNumberForKeywords
+
+The provided VersionNumber '%s' is invalid. Provide an integer value or use the keyword '%s' or '%s' for the build number.
+
+# errorInvalidBuildNumber
+
+The provided VersionNumber '%s' is invalid. Provide an integer value or use the keyword '%s' for the build number.
